REF: Moved async utils to separate module
